// website/src/types/appwrite.d.ts
import { type Models } from "node-appwrite";

// This file is auto-generated by the Appwrite CLI.
// You can regenerate it by running `appwrite types -l ts src/types`.

export type Links = Models.Row & {
  title: string;
  url: string;
  icon: string;
  order: number;
  active: boolean;
  type: string;
  category: string;
};